requirements.txt
cffi==2.0.0
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
fonttools==4.62.0
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pillow==12.1.1
pycparser==3.0
pycryptodome==3.23.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
